requirements-dev.txt
# Thalos Prime v1.0 - Development Dependencies
#
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# Development Tools
pre-commit>=3.0.0
tox>=4.0.0

# Debugging
ipython>=8.0.0
pdb++>=0.10.0

# Performance Testing
locust>=2.15.0
pytest-benchmark>=4.0.0

# Security Testing
bandit>=1.7.0
safety>=2.3.0
